+++ test/pseudoace/test_cli.clj
@@ -1,8 +1,8 @@
-(ns pseudoace.test-core
+(ns pseudoace.test-cli
   (:require
    [clojure.java.io :as io]
    [clojure.test :refer (deftest is)]
-   [pseudoace.core :as core]
+   [pseudoace.cli :as cli]
    [pseudoace.utils :as utils]))
 
 (deftest test-move-helper-log-file
@@ -10,14 +10,14 @@
                             (System/getProperty "java.io.tmpdir")
                             "dummy-edn-log-dir")]
     (.mkdir dummy-log-dir)
-    (let [dummy-helper-file (io/file dummy-log-dir core/helper-filename)]
+    (let [dummy-helper-file (io/file dummy-log-dir cli/helper-filename)]
       (utils/with-outfile dummy-helper-file
         (println "dummy help edn content"))
       ;; Call function under test
-      (core/move-helper-log-file (.getParent dummy-helper-file))
+      (cli/move-helper-log-file (.getParent dummy-helper-file))
       ;; Check expectations
       (let [dummy-log-dir (.getParent dummy-helper-file)
-            expected-dest-file (core/helper-dest-file dummy-log-dir)]
+            expected-dest-file (cli/helper-dest-file dummy-log-dir)]
         (is (.exists expected-dest-file))
         (is (not (.exists dummy-helper-file)))
         (io/delete-file expected-dest-file)