+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Change-log
 
+## 0.4.15 (unreleased)
+  - Added generated schema for WS257.
+  - Make output of generated schema valid EDN.
+  - Use latest version of datomic 0.9.5554.
+
 ## 0.4.14 (2016-11-11)
   - Allow migrations with no locatable schema and/or schema fixups.